refactor(seo): consolidate JSON-LD modules into json-ld.ts
- Merge blog-posting.ts, site-json-ld.ts into json-ld.ts
- Add JsonLd<T> generic type for type-safe JSON-LD structures
- Rename buildWebSiteJsonLd to webSiteJsonLd for consistency
- Consolidate tests into json-ld.test.ts
- Update imports in app/page.tsx and app/blog/[slug]/page.tsx
