+++ app/blog/[slug]/page.tsx
@@ -2,8 +2,7 @@ import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { Badge } from "@/components/ui/badge";
 import { listBlogSlugs, loadBlogPost } from "@/lib/content/blog/mdx";
-import { blogPostingJsonLd } from "@/lib/metadata/blog-posting";
-import { jsonLdScript } from "@/lib/metadata/json-ld";
+import { blogPostingJsonLd, jsonLdScript } from "@/lib/metadata/json-ld";
 import { pageOpenGraph, pageTwitter } from "@/lib/metadata/page";
 import { config } from "@/lib/site/config";
 import { blogPostOgImagePath } from "@/lib/site/routes";
